Move boxview tests to layout tests
This is really a Grid test
